+++ assets/javascripts/kitten/components/navigation/nav-bar/test.js
@@ -1,32 +1,32 @@
 import React from 'react'
 import renderer from 'react-test-renderer'
 import 'jest-styled-components'
-import { TabBar } from './next'
+import { NavBar } from './index'
 
-describe('<TabBar />', () => {
+describe('<NavBar />', () => {
   let component
 
-  describe('simple TabBar with zIndex set to 105', () => {
+  describe('simple NavBar with zIndex set to 105', () => {
     beforeEach(() => {
       component = renderer
         .create(
-          <TabBar
+          <NavBar
             className="customClass"
             listProps={{ className: 'customListClass' }}
             navProps={{ className: 'customNavClass' }}
           >
-            <TabBar.ListItem linkProps={{ href: '#1' }} active>
+            <NavBar.ListItem linkProps={{ href: '#1' }} active>
               Item 1
-            </TabBar.ListItem>
-            <TabBar.ListItem
+            </NavBar.ListItem>
+            <NavBar.ListItem
               className="customItemClass"
               linkProps={{ href: '#2', className: 'customLinkClass' }}
             >
               Item 2
-            </TabBar.ListItem>
-            <TabBar.ListItem linkProps={{ href: '#3' }}>Item 3</TabBar.ListItem>
-            <TabBar.ListItem linkProps={{ href: '#4' }}>Item 4</TabBar.ListItem>
-          </TabBar>,
+            </NavBar.ListItem>
+            <NavBar.ListItem linkProps={{ href: '#3' }}>Item 3</NavBar.ListItem>
+            <NavBar.ListItem linkProps={{ href: '#4' }}>Item 4</NavBar.ListItem>
+          </NavBar>,
         )
         .toJSON()
     })