+++ assets/javascripts/kitten/next/index.js
@@ -1,4 +1,3 @@
 export { Modal } from '../components/modals/next'
 export { Button } from '../components/buttons/button/button'
 export { Carousel } from '../components/carousel/carousel/carousel'
-export { TabBar } from '../components/tab-bar/next'