+++ CHANGELOG.md
@@ -4,8 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: Add `TabBar` Next component.
-- Fix: Adjust styles for legacy `TabBar`.
+- Feature: Add `NavBar` component.
 
 ## [2.69.0] - 2020-05-12